Router State and Entity Composition
