+++ src/products/store/reducers/index.ts
@@ -1,5 +1,5 @@
-import { ActionReducerMap, createFeatureSelector, createSelector } from '@ngrx/store';
-import * as fromPizzas from './pizzas.reducer';
+import { ActionReducerMap, createFeatureSelector } from '@ngrx/store';
+import * as fromPizzas from './pizzas.reducers';
 
 export interface ProductsState {
   pizzas: fromPizzas.PizzaState;
@@ -12,17 +12,3 @@ export const reducers: ActionReducerMap<ProductsState> = {
 export const getProductsState = createFeatureSelector<ProductsState>(
   'products'
 );
-
-// pizzas state
-export const getPizzaState = createSelector(
-  getProductsState,
-  (state: ProductsState) => state.pizzas
-);
-
-export const getPizzasEntities = createSelector(getPizzaState, fromPizzas.getPizzasEntities);
-export const getAllPizzas = createSelector(
-  getPizzasEntities,
-  (entities) => Object.keys(entities).map(id => entities[parseInt(id, 10)])
-);
-export const getPizzasLoaded = createSelector(getPizzaState, fromPizzas.getPizzasLoaded);
-export const getPizzasLoading = createSelector(getPizzaState, fromPizzas.getPizzasLoading);
