+++ src/products/store/selectors/index.ts
@@ -1,0 +1,1 @@
+export * from './pizzas.selectors';
